requirements.txt
flask==3.0.2
boto3==1.34.131
botocore==1.34.131
azure-storage-blob==12.20.0
azure-identity==1.17.1
python-dotenv==1.0.0